requirements.txt
torch==1.8.1+cu111
torchvision==0.9.1+cu111
opencv-python
tqdm
attrdict
scipy
pycocotools
hydra-core
hydra_colorlog
shapely
matplotlib>=3.2.2
scipy>=1.4.1
Pillow
requests
seaborn
